Enable Secure Boot in all required recipes
In the target, Secure Boot starts from the firmware (u-boot), adds the
signing keys, and verifies the bootloader (systemd-boot) and kernel
(Linux).

The gen-sbkeys recipe generates keys in the host during build time. If
keys are found, it will skip the keys generation. As long as the keys
are the same, it will guarantee build reproducibility. 

sbsign bbclass is used to sign the binaries. As Mikko's said:

    sbsign is the name of the tool used to sign these binaries. Hence
    the name of this class to sbsign and variables with SBSIGN prefix.

Signed-off-by: Javier Tia <javier.tia@linaro.org>